List of messageServices to allow different ways of sending; refactoring
