+++ build.gradle
@@ -29,7 +29,7 @@ dependencies {
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign:2.2.5.RELEASE'
     implementation 'org.springframework.cloud:spring-cloud-dependencies:Hoxton.SR8'
 
-    compileOnly 'org.projectlombok:lombok'
+    implementation 'org.projectlombok:lombok'
     runtimeOnly 'com.h2database:h2'
     annotationProcessor 'org.projectlombok:lombok'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
