+++ src/main/java/com/phillippko/analytics/client/FeignMessageClient.java
@@ -1,15 +1,16 @@
 package com.phillippko.analytics.client;
 
 import com.phillippko.analytics.dto.MessageOutgoingDto;
+import feign.Response;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import java.net.URI;
 
 
-@FeignClient(name = "messages", url = "placeholder")
-public interface MessageClient {
+@FeignClient(name = "messages", url = "")
+public interface FeignMessageClient {
 
     @PostMapping("")
-    void send(URI baseUrl, MessageOutgoingDto message);
+    Response send(URI baseUrl, MessageOutgoingDto message);
 }
