+++ src/main/java/com/phillippko/analytics/controller/TemplateController.java
@@ -2,29 +2,21 @@ package com.phillippko.analytics.controller;
 
 import com.phillippko.analytics.dto.MessageIncomingDto;
 import com.phillippko.analytics.dto.TemplateDto;
-import com.phillippko.analytics.service.MessageService;
 import com.phillippko.analytics.service.TemplateService;
+import com.phillippko.analytics.service.messageSenders.MessageService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.ApplicationContext;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
 public class TemplateController {
     private final TemplateService templateService;
-    private MessageService messageService;
+    private List<MessageService> messageServices;
 
-    @Value("${message-service:feign-message-service}")
-    private String messageServiceType;
-
-    @Autowired
-    void setMessageService(ApplicationContext context) {
-        messageService = (MessageService) context.getBean(messageServiceType);
-    }
 
     @PostMapping("template")
     void postMapping(@RequestBody TemplateDto templateDto) {
@@ -33,6 +25,6 @@ public class TemplateController {
 
     @PostMapping("send-message")
     void sendMessage(@RequestBody MessageIncomingDto messageIncomingDto) {
-        messageService.sendMessage(messageIncomingDto);
+        messageServices.forEach(messageService -> messageService.sendMessage(messageIncomingDto));
     }
 }
