--- src/main/java/com/phillippko/analytics/utils/MessageVariable.java
@@ -1,10 +1,0 @@
-package com.phillippko.analytics.utils;
-
-public class MessageVariable {
-    public String value;
-
-    public MessageVariable(String name, String value) {
-
-        this.value = value;
-    }
-}
